added elements and config finder.
