+++ packages/cli/src/commands/init.ts
@@ -1,8 +1,6 @@
 import { Command, flags } from '@oclif/command'
-import * as path from 'path'
 import { promptUser, getListOptions } from '../utils/commands'
-var execa = require('execa')
-var Listr = require('listr')
+import { questions } from '../questions/init'
 var yeoman = require('yeoman-environment')
 var env = yeoman.createEnv()
 env.register(require.resolve('@wcfactory/generator-wcfactory/generators/init'), 'wcfactory:init')
@@ -33,100 +31,7 @@ export default class Init extends Command {
     flags = await promptUser(questions, flags, this)
     // add a year
     flags.year = new Date().getFullYear()
-    // now we are going to assemble a task list
-    const listOptions = getListOptions(flags)
-    // assemble a list
-    const list = new Listr([
-      {
-        title: 'Setting up files',
-        task: async () => {
-          const res = await env.run('wcfactory:init', flags)
-        }
-      },
-      {
-        title: 'Initializing Git',
-        task: async () => execa('git', ['init']).stdout
-      },
-      {
-        title: 'Installing dependencies',
-        task: async () => execa('yarn', ['install']).stdout
-      },
-      {
-        title: 'Running Initalizer',
-        task: async () => execa('yarn', ['run', 'init']).stdout
-      },
-      {
-        title: 'Rebuilding cache',
-        task: async () => execa('yarn', ['run', 'rebuild-wcfacache']).stdout
-      }
-    ], listOptions)
-    // run the list
-    list.run();
+    // kick off generator
+    env.run('wcfactory:init', flags)
   }
 }
-
-
-/**
- * A listing of all prompt questions
- */
-const questions: any = [
-  {
-    type: "string",
-    name: "humanName",
-    message: "Name of this factory",
-    required: true,
-    store: true,
-    default: (flags: any) => {
-      return path.basename(process.cwd())
-    }
-  },
-  {
-    type: "string",
-    name: "description",
-    message: "Description",
-    required: true,
-    store: true
-  },
-  {
-    type: "string",
-    name: "name",
-    message: "Repo name (a valid git / npm machine name)",
-    required: true,
-    default: (flags: any) => flags.humanName,
-    validate: (value: any) => {
-      if ((/([a-z]*)-([a-z]*)/).test(value)) { return true; }
-      return 'name requires a hyphen and all lowercase';
-    }
-  },
-  {
-    type: "string",
-    name: "orgNpm",
-    message: "NPM organization name (include @)",
-    required: true,
-    store: true,
-    default: (flags: any) => '@' + flags.name,
-    validate: (value: any) => {
-      if (!(/ /gm).test(value)) { return true; }
-      return 'can not contain a space';
-    }
-  },
-  {
-    type: "string",
-    name: "orgGit",
-    message: "Git organization name",
-    required: true,
-    store: true,
-    default: (flags: any) => flags.name.replace('@', ''),
-    validate: (value: any) => {
-      if (!(/ /).test(value)) { return true; }
-      return 'can not contain a space';
-    }
-  },
-  {
-    type: "string",
-    name: "gitRepo",
-    message: "Git repo (full git address)",
-    required: true,
-    default: (flags: any) => `git@github.com:${flags.orgGit}/${flags.name}.git`
-  }
-]
